+++ src/gui/tools/select_tool.cpp
@@ -29,7 +29,7 @@ public:
     QIcon icon() const override { return QIcon::fromTheme("edit-select"); }
     QString name() const override { return i18n("Select"); }
     QString action_name() const override { return QStringLiteral("tool_select"); }
-    QKeySequence key_sequence() const override { return Qt::Key_F3; }
+    QKeySequence key_sequence() const override { return Qt::Key_F1; }
     static int static_group() noexcept { return Registry::Core; }
     int group() const noexcept override { return static_group(); }